useCreateGroup hook for reusability
